Django/eve/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extensions==0.7.1
gunicorn==20.1.0
mysqlclient==2.1.1
psycopg2-binary==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3
tomli==2.0.1
